% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES.R
\name{ES}
\alias{ES}
\title{Calculates ES the Hurlbert diversity index}
\usage{
ES(
  df,
  by = NULL,
  var_species = "species",
  var_count = "count",
  check_species = TRUE,
  df_species = NULL,
  n = 100
)
}
\arguments{
\item{df}{a dataframe of species observations}

\item{by}{a vector of column names found in \code{df} by which calculations
should be grouped \emph{e.g. c("station","date")}}

\item{var_species}{name of the column in \code{df} containing species names}

\item{var_count}{name of the column in \code{df} containing count/density/abundance}

\item{check_species}{boolean, default = TRUE. If TRUE, then only species found
in the species list are included in \emph{H'} index. By default,
the AZTI species list is used.}

\item{df_species}{\emph{optional} dataframe with user-specified species list.}

\item{n}{subsample size, usually indicated in the abbreviated name for the index.
e.g. \code{ES100} is ES calculated with a subsample of 100 individuals
(the default) The sampling is done for each \emph{by} group.}
}
\value{
a list of two dataframes:
\itemize{
\item \code{ES} : results of the ES index calculations. For each unique
combination of \code{by}variables the following values are calculated:
\itemize{
\item \verb{ES<n>} : the expected sample species richness, \emph{ES}
for a sample size \code{n} of 100, the variable name is \code{ES100}
\item \code{S} : the number of species
\item \code{N} : the number of individuals
}
\item \code{match} : the original dataframe with columns added from the species list.
For a user-specified list provided \code{df_species}, all columns will be included.
If the user-specified species list contains only a single column with species
names, then a new column \code{match} will be created, with a value of \code{1} indicating
a match and an \code{NA} value where no match was found.
}

For the default AZTI species list the following additional columns will be
included:
\itemize{
\item \code{group} : showing the AMBI species group
\item \code{RA} : indicating that the species is \emph{reallocatable} according to the
AZTI list. That is, it could be re-assigned to a different species group.
}
}
\description{
\code{\link[=ES]{ES()}} matches a list of species counts with the AMBI species list
and calculates \emph{ES} the Hurlbert diversity index, which gives the
expected sample species richness for a subsample of the population.
\href{#references}{(Hurlbert, 1971)}
}
\details{
If the function is called with the argument \code{check_species = TRUE} then
only species which are successfully matched with the specified species
list are included in the calculations. This is the default. If the function
is called with \code{check_species = FALSE}then all rows are counted.
}
\examples{

ES(test_data)

}
\references{
Hurlbert, S.H. (1971), The Nonconcept of Species Diversity: A Critique and Alternative Parameters. Ecology, 52: 577-586.
\doi{doi:10.2307/1934145}
}
\seealso{
For more details, see`vignette("other-indices").
}
