% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO_species.R
\name{NO_species}
\alias{NO_species}
\title{Returns species list used by Norwegian index calculations}
\usage{
NO_species()
}
\value{
A data frame with 760 rows* and 10 columns:

\describe{
\item{species}{Species name or genus (spp.)}
\item{ES100avg_2018}{Species sensitivity value for ES100 used to calculate \code{\link[=NSI]{NSI()}} version 2018}
\item{ES100min5_2018}{Species sensitivity value used to calculate \code{\link[=ISI]{ISI()}} version 2018}
\item{ES100avg_2012}{Species sensitivity value for ES100 used to calculate \code{\link[=NSI]{NSI()}} version 2012}
\item{ES100min5_2012}{Species sensitivity value used to calculate \code{\link[=ISI]{ISI()}} version 2012}
\item{group_AMBI}{AMBI species group. This information is included by
\href{#references}{Borgersen et al. (2019)}. For the
authoritative list see \code{\link[=AMBI_species]{AMBI_species()}}}
\item{group_NSI}{NSI species group. See \href{#references}{Borgersen et al. (2019)}}
\item{group_ISI}{ISI species group}
\item{Revidert}{Indicates that information has been revised.}
\item{Kommentar}{Comment in Norwegian describing revision.}
}
}
\description{
\code{\link[=NO_species]{NO_species()}} returns a dataframe with list of species and sensitivities for
use in functions calculating Norwegian indices \code{\link[=ISI]{ISI()}} and \code{\link[=NSI]{NSI()}}.

\emph{latest version 13th November 2024}
}
\details{
This table contains the species-specific sensitivities used to calculate the
Norwegian indices \code{\link[=ISI]{ISI()}} and \code{\link[=NSI]{NSI()}}.

Similarly to the \code{AMBI} index, the principle of the Norwegian quality indices
is that high species diversity indicates good environmental status and species
primarily found at sites with high diversity and considered sensitive, while
sites with low species diversity are considered as stressed environments and
species more frequently found at such sites are considered more tolerant.
Species diversity thus serves as a proxy for the degree of disturbance at a site.

The calculation of species sensitivity values for ISI and NSI is based on
analysis of \code{ES100} values calculated for a large number of grab samples from
the Norwegian monitoring data. The sensitivity values for a species are
the weighted averages of \code{ES100} values across samples where the species
was present.

\itemize{
\item \emph{ES100avg_2018} \cr The weighted average of \code{ES100} values across
samples where the species was present. Theses are the updated 2018 values.\cr
\item \emph{ES100min5_2018} \cr The weighted average of \code{ES100} values across
samples where the species was present but based on the 5 samples having the
lowest \code{ES100} value. This value was determined only for species found in
at least six samples. Theses are the updated 2018 values.\cr
\item \emph{ES100avg_2012} \cr The weighted average of \code{ES100} values across
samples where the species was present.  These are the values determined
by \href{#references}{Rygg & Norling (2013)} \cr
\item \emph{ES100min5_2012} \cr The weighted average of \code{ES100} values across
samples where the species was present but based on the 5 samples having the
lowest \code{ES100} value. This value was determined only for species found in
at least six samples. These are the values determined
by \href{#references}{Rygg & Norling (2013)} \cr
}
,
For more details, please see \href{#references}{Borgersen et al. (2019)}:
}
\examples{

NO_species() \%>\% head()

NO_species() \%>\% tail()

}
\references{
"Sensitivitetsverdier for NSI2018 og ISI 2018_13.11.2024.xlsx"
retrieved 01-09-2026
https://www.vannportalen.no/sharepoint/downloaditem?id=01FM3LD2WM4F42XCAEDNBZLJRZMUOM66OP
Borgersen, G., Trannum, H.C., Gundersen, H., Vedal, J. (2019). Oppdatering av bløtbunnsartenes sensitivitetsverdier (Report No. 7366). Norsk institutt for vannforskning (NIVA).
\url{https://hdl.handle.net/11250/2600903}
Rygg, B., Norling, K. (2013). Norwegian Sensitivity Index (NSI) for marine macroinvertebrates, and an update of Indicator Species Index (ISI). (Report No. 6475). Norsk institutt for vannforskning (NIVA). 46 pp.
}
\seealso{
\code{\link[=NSI]{NSI()}} and \code{\link[=ISI]{ISI()}} use this species list to calculate these indices.
}
