% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NQI1.R
\name{NQI1}
\alias{NQI1}
\title{Calculates NQI1, the Norwegian Quality Index}
\usage{
NQI1(AMBI, N, S)
}
\arguments{
\item{AMBI}{AMBI, the AZTI Marine Biotic Index, calculated using \code{\link[=AMBI]{AMBI()}}}

\item{N}{number of individuals - returned by both \code{\link[=AMBI]{AMBI()}} and \code{\link[=Hdash]{Hdash()}}}

\item{S}{number of species - returned by both \code{\link[=AMBI]{AMBI()}} and \code{\link[=Hdash]{Hdash()}}}
}
\value{
\code{NQI1} index value
}
\description{
\code{\link[=NQI1]{NQI1()}} calculates the Norwegian Quality Index, NQI1 \href{#references}{(Rygg, 2006)}

The NQI1 is based on AMBI and can only be calculated after first calculating
\emph{AMBI}, the AZTI Marine Biotic Index, using output from the function \code{\link[=AMBI]{AMBI()}}.

NQI1 is the only Norwegian benthic fauna index which is inter-calibrated with
all countries belonging to the North East Atlantic Geographical
Intercalibration Group (NEAGIG)\href{#references}{(Borgersen et al, 2019)} .
}
\details{
While \code{\link[=AMBI]{AMBI()}} takes a dataframe of observations as an argument,
\code{\link[=NQI1]{NQI1()}} does \emph{not}. Instead, similarly to the DKI functions, it
takes values of the input parameters, either single values or as vectors.

To calculate NQI1 for a dataframe of \code{AMBI} values, it could be called from
e.g. within a \code{\link[dplyr:mutate]{dplyr::mutate()}} function call. See the examples below.
}
\examples{

# Simple example

NQI1(AMBI = 1.61, N = 25, S = 6)


# ------ Example workflow for calculating NQI1 from species counts ----

# calculate AMBI index using Norwegian example data

df <- AMBI(test_data_NO,
               by = "station",
               var_rep = "sample",
               var_species = "species",
               var_count = "count",
               quiet = TRUE)[["AMBI"]]

# show AMBI results
df

# calculate NQI1 from AMBI results
df <- dplyr::mutate(df, NQI1 = NQI1(AMBI, N, S))
dplyr::select(df, AMBI, N, S, NQI1)

}
\references{
Rygg, B. (2006) Developing indices for quality-status classification of marine soft-bottom fauna in Norway. (Report No. 5208) Norwegian Institute for Water Research (NIVA). 33 pp.
\url{https://hdl.handle.net/11250/213219}
Borgersen, G., Trannum, H.C., Gundersen, H., Vedal, J. (2019). Oppdatering av bløtbunnsartenes sensitivitetsverdier (Report No. 7366). Norsk institutt for vannforskning (NIVA).
\url{https://hdl.handle.net/11250/2600903}
}
\seealso{
For more details, see`vignette("other-indices").
}
