% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSI.R
\name{NSI}
\alias{NSI}
\title{Calculates NSI, the Norwegian Sensitivity Index}
\usage{
NSI(
  df,
  by = NULL,
  var_species = "species",
  var_count = "count",
  df_species = NULL,
  version = "2018"
)
}
\arguments{
\item{df}{a dataframe of species observations}

\item{by}{a vector of column names found in \code{df} by which calculations
should be grouped \emph{e.g. c("station","date")}}

\item{var_species}{name of the column in \code{df} containing species names}

\item{var_count}{name of the column in \code{df} containing count/density/abundance}

\item{df_species}{\emph{optional} dataframe with user-specified species list.}

\item{version}{\emph{string}, version of the index to be calculated.
The default value is \code{"2018"}. The the only other valid
value for \code{version} is \code{"2012"}.}
}
\value{
a list of two dataframes:
\itemize{
\item \code{NSI} : results of the NSI index calculations. For each unique
combination of \code{by}variables the following values are calculated:
\itemize{
\item \verb{NSI<version>} : the sensitivity index \code{NSI2018} or \code{NSI2012}
\item \code{S} : the number of species
\item \code{N} : the number of individuals
}
\item \code{match} : the original dataframe with columns added from the species list.
For a user-specified list provided \code{df_species}, all columns will be included.
If the user-specified species list contains only a single column with species
names, then a new column \code{match} will be created, with a value of \code{1} indicating
a match and an \code{NA} value where no match was found.
One of the following columns will be included, depending on the NSI version:
\itemize{
\item \code{ES100avg_2018} : species sensitivity values used to calculate \code{NSI2018}
\item \code{ES100avg_2012} : species sensitivity values used to calculate \code{NSI2012}
}
}
}
\description{
\code{\link[=NSI]{NSI()}} matches a list of species counts with the Norwegian species list
and calculates \emph{NSI} the Norwegian Sensitivity Index.
\href{#references}{(Borgersen et al, 2019)}
}
\details{
If the function is called with the argument \code{check_species = TRUE} then
only species which are successfully matched with the specified species
list are included in the calculations. This is the default. If the function
is called with \code{check_species = FALSE}then all rows are counted.
}
\examples{

NSI(test_data_NO, by="station", var_species ="species", var_count="count")

}
\references{
Borgersen, G., Trannum, H.C., Gundersen, H., Vedal, J. (2019). Oppdatering av bløtbunnsartenes sensitivitetsverdier (Report No. 7366). Norsk institutt for vannforskning (NIVA).
\url{https://hdl.handle.net/11250/2600903}

Rygg, B., Norling, K. (2013). Norwegian Sensitivity Index (NSI) for marine macroinvertebrates, and an update of Indicator Species Index (ISI). (Report No. 6475). Norsk institutt for vannforskning (NIVA). 46 pp.
}
\seealso{
\code{\link[=ISI]{ISI()}} which calculates the Indicator Species Index, another Norwegian
benthic diversity index.
For more details, see`vignette("other-indices").
}
