% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_DK.R
\docType{data}
\name{test_data_DK}
\alias{test_data_DK}
\title{Danish test dataset}
\format{
The test dataset \code{test_data_DK} has 112 rows and 4 variables:
\describe{
\item{station}{2 sampling sites "42", "49" \emph{(ObservationsStedNavn)}}
\item{sample}{unique samples taken at each site,
identified \emph{1}, \emph{2}, \emph{3}, \emph{4}, \emph{5} \emph{(Prøvetagningsnummer)}}
\item{species}{Name of observed species/taxon \emph{(Artsnavn)}}
\item{count}{Number of individuals \emph{(Antal)}}
}
}
\source{
\href{https://odaforalle.au.dk}{Danish National Aquatic Monitoring and Assessment Programme.}
}
\usage{
test_data_DK
}
\description{
Example data from Danish monitoring data. Used for testing DKI calculations.
}
\examples{
head(test_data_DK)

}
\keyword{datasets}
