% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_NO.R
\docType{data}
\name{test_data_NO}
\alias{test_data_NO}
\title{Norwegian test dataset}
\format{
The test dataset \code{test_data_NO} has 396 rows and 4 variables:
\describe{
\item{station}{2 sampling sites "Ærøy (U10)", "Ærøydypet (U12)" \emph{(Vannlokalitetsnavn)}}
\item{sample}{unique samples taken at each site,
identified \emph{G1}, \emph{G2} \emph{(Provenr)}}
\item{species}{Name of observed species/taxon \emph{(VitenskapligNavn)}}
\item{count}{Number of individuals \emph{(Verdi)}}
}
}
\source{
\href{https://vann-nett.no/}{Vann-Nett}
}
\usage{
test_data_NO
}
\description{
Example data from Norwegian monitoring data. Used for testing Norwegian index calculations.
}
\examples{
head(test_data_NO)

}
\keyword{datasets}
