% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBIr-package.R
\docType{package}
\name{ambiR-package}
\alias{ambiR}
\alias{ambiR-package}
\title{ambiR: Calculate AZTI’s Marine Biotic Index}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate AZTI’s Marine Biotic Index _AMBI_, as well as the Shannon Diversity Index _H'_ and the Danish benthic fauna quality index _DKI_. _Borja, A., J. Franco, V. Pérez, 2000. A marine biotic index to establish the ecological quality of soft bottom benthos within European estuarine and coastal environments Marine Pollution Bulletin 40(12) 1100-1114_.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://niva-denmark.github.io/ambiR/}
  \item \url{https://github.com/niva-denmark/ambiR/}
  \item Report bugs at \url{https://github.com/niva-denmark/ambiR/issues}
}

}
\author{
\strong{Maintainer}: Ciarán J. Murray \email{cjm@niva-dk.dk} (\href{https://orcid.org/0000-0003-0260-2008}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ángel Borja \email{aborja@azti.es} (\href{https://orcid.org/0000-0003-1601-2025}{ORCID})
  \item Sarai Pouso \email{spouso@azti.es} (\href{https://orcid.org/0000-0003-3208-9882}{ORCID})
  \item Iñigo Muxika \email{imuxika@azti.es} (\href{https://orcid.org/0000-0002-9181-3781}{ORCID})
  \item Joxe Mikel Garmendia \email{jgarmendia@azti.es} (\href{https://orcid.org/0000-0002-9403-1777}{ORCID})
}

Other contributors:
\itemize{
  \item Steen Knudsen \email{steen.knudsen@niva-dk.dk} (\href{https://orcid.org/0000-0003-0428-9940}{ORCID}) [contributor]
}

}
\keyword{internal}
