% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMBIr-package.R
\docType{package}
\name{ambiR-package}
\alias{ambiR}
\alias{ambiR-package}
\title{ambiR: Calculate AZTI’s Marine Biotic Index}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate AZTI’s Marine Biotic Index _AMBI_. The included list of benthic fauna species according to their sensitivity to pollution. Matching species in sample data to the list allows the calculation of fractions of individuals in the different sensitivity categories and thereafter the AMBI index. The Shannon Diversity Index _H'_ and the Danish benthic fauna quality index _DKI_ can also be calculated, as well as the multivariate _M-AMBI_ index. _Borja, A., J. Franco, V. Pérez, 2000. A marine biotic index to establish the ecological quality of soft bottom benthos within European estuarine and coastal environments Marine Pollution Bulletin 40(12) 1100-1114_.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://niva-denmark.github.io/ambiR/}
  \item \url{https://github.com/niva-denmark/ambiR/}
  \item Report bugs at \url{https://github.com/niva-denmark/ambiR/issues}
}

}
\author{
\strong{Maintainer}: Ciarán J. Murray \email{cjm@niva-dk.dk} (\href{https://orcid.org/0000-0003-0260-2008}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ángel Borja \email{aborja@azti.es} (\href{https://orcid.org/0000-0003-1601-2025}{ORCID})
  \item Sarai Pouso \email{spouso@azti.es} (\href{https://orcid.org/0000-0003-3208-9882}{ORCID})
  \item Iñigo Muxika \email{imuxika@azti.es} (\href{https://orcid.org/0000-0002-9181-3781}{ORCID})
  \item Joxe Mikel Garmendia \email{jgarmendia@azti.es} (\href{https://orcid.org/0000-0002-9403-1777}{ORCID})
}

Other contributors:
\itemize{
  \item Steen Knudsen \email{steen.knudsen@niva-dk.dk} (\href{https://orcid.org/0000-0003-0428-9940}{ORCID}) [contributor]
  \item GES4SEAS (Grant Agreement 101059877 - GES4SEAS. The GES4SEAS project has been approved under the HORIZON-CL6-2021-BIODIV-01-04 call: 'Assess and predict integrated impacts of cumulative direct and indirect stressors on coastal and marine biodiversity, ecosystems and their services'. Funded by the European Union. Views and opinions expressed are however those of the authors only and do not necessarily reflect those of the European Union or UK Research and Innovation. Neither the European Union nor the granting authority can be held responsible for them.) [funder]
}

}
\keyword{internal}
