% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H_sal}
\alias{H_sal}
\title{Maximum H' as a linear function of salinity}
\source{
\url{https://dce2.au.dk/pub/SR93.pdf}
}
\usage{
H_sal(psal, intercept = 2.117, slope = 0.086)
}
\arguments{
\item{psal}{numeric salinity}

\item{intercept}{numeric, default 2.117}

\item{slope}{numeric default 0.086}
}
\value{
a numeric value H_max
}
\description{
\emph{AMBI_sal} and \emph{H_sal} are called \emph{AMBI_min} and
\emph{H_max} in the DKI documentation
They are renamed to reflect the fact that they are
functions of salinity and not min/max  values from
the data being used in the function
}
\examples{
H_sal(20.1)

}
